change component name prefix adjectives to suffix
fixes #201
