+++ src/components/CheckboxLabeled/CheckboxLabeled.less
@@ -1,0 +1,46 @@
+.lucid-CheckboxLabeled {
+	align-items: center;
+	color: @color-darkGray;
+	cursor: pointer;
+	display: flex;
+	font: @font;
+	justify-content: flex-start;
+
+	> .lucid-Checkbox {
+		flex: 0 0 @Checkbox-size;
+		margin-right: @size-XXS;
+	}
+
+	&&-is-disabled {
+		color: @color-mediumGray;
+		cursor: not-allowed;
+	}
+
+	&:hover .lucid-Checkbox-visualization-container {
+		border-color: @color-primary;
+		transition: @Checkbox-animation-timing;
+	}
+
+	&:hover .lucid-Checkbox-visualization-glow {
+		.opacity(0.3);
+		transform: @Checkbox-glow-scale;
+	}
+
+	&&-is-disabled .lucid-Checkbox-visualization-container,
+	&&-is-disabled:hover .lucid-Checkbox-visualization-container {
+		background-color: @Checkbox-disabled-backgroundColor;
+		border-color: @Checkbox-disabled-borderColor;
+	}
+
+	&&-is-disabled:hover .lucid-Checkbox-visualization-glow {
+		.opacity(0);
+		transform: none;
+	}
+
+
+	&&-is-disabled&&-is-selected:hover .lucid-Checkbox-visualization-container,
+	&&-is-selected .lucid-Checkbox-visualization-container {
+		background-color: @color-primary;
+		border-color: @color-primary;
+	}
+}
