--- src/components/LabeledRadioButton/examples/label-as-child.jsx
@@ -1,25 +1,0 @@
-import React from 'react';
-import { LabeledRadioButton } from '../../../index';
-
-const style = {
-	marginBottom: '3px'
-};
-
-export default React.createClass({
-	render() {
-		return (
-			<section>
-				<section>
-					<LabeledRadioButton style={style}>
-						<LabeledRadioButton.Label>Just text</LabeledRadioButton.Label>
-					</LabeledRadioButton>
-					<LabeledRadioButton style={style}>
-						<LabeledRadioButton.Label>
-							<span>HTML element</span>
-						</LabeledRadioButton.Label>
-					</LabeledRadioButton>
-				</section>
-			</section>
-		);
-	}
-});
