+++ src/components/RadioGroup/RadioGroup.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { lucidClassNames } from '../../util/style-helpers';
 import { createClass } from '../../util/component-definition';
 import getRandom from '../../util/random';
-import LabeledRadioButton from '../LabeledRadioButton/LabeledRadioButton';
+import RadioButtonLabeled from '../RadioButtonLabeled/RadioButtonLabeled';
 import RadioButton from '../RadioButton/RadioButton';
 import reducers from './RadioGroup.reducers';
 
@@ -116,7 +116,7 @@ const RadioGroup = createClass({
 			>
 				{_.map(radioButtonChildProps, (radioButtonChildProp, index) => {
 					return (
-						<LabeledRadioButton
+						<RadioButtonLabeled
 								{...radioButtonChildProp}
 								isSelected={actualSelectedIndex === index}
 								key={index}
