+++ src/components/TextFieldValidated/TextFieldValidated.less
@@ -1,0 +1,10 @@
+.lucid-TextFieldValidated {
+	width: 100%;
+
+	&.lucid-Validation-is-error {
+		.lucid-TextField {
+			.inset-box-shadow(@featured-color-danger);
+			border-color: @featured-color-danger;
+		}
+	}
+}
