+++ src/components/TextFieldValidated/TextFieldValidated.spec.jsx
@@ -1,0 +1,7 @@
+import { common } from '../../util/generic-tests';
+import TextFieldValidated from './TextFieldValidated';
+
+describe('TextFieldValidated', () => {
+	common(TextFieldValidated);
+});
+
