+++ src/components/TextFieldValidated/examples/basic.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { TextFieldValidated } from '../../../index';
+
+export default React.createClass({
+	render() {
+		return <TextFieldValidated Error='Nope, not even close!' />;
+	}
+});
