--- src/components/ValidatedTextField/ValidatedTextField.less
@@ -1,10 +1,0 @@
-.lucid-ValidatedTextField {
-	width: 100%;
-
-	&.lucid-Validation-is-error {
-		.lucid-TextField {
-			.inset-box-shadow(@featured-color-danger);
-			border-color: @featured-color-danger;
-		}
-	}
-}
