--- src/components/ValidatedTextField/ValidatedTextField.spec.jsx
@@ -1,7 +1,0 @@
-import { common } from '../../util/generic-tests';
-import ValidatedTextField from './ValidatedTextField';
-
-describe('ValidatedTextField', () => {
-	common(ValidatedTextField);
-});
-
