--- src/components/ValidatedTextField/examples/basic.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { ValidatedTextField } from '../../../index';
-
-export default React.createClass({
-	render() {
-		return <ValidatedTextField Error='Nope, not even close!' />;
-	}
-});
