+++ src/styles/components.less
@@ -20,9 +20,9 @@
 @import '../components/Icon/EligibilityIcon/EligibilityIcon';
 @import '../components/Icon/WarningIcon/WarningIcon';
 @import '../components/Icon/ResizeIcon/ResizeIcon';
-@import '../components/LabeledCheckbox/LabeledCheckbox';
-@import '../components/LabeledRadioButton/LabeledRadioButton';
-@import '../components/LabeledSwitch/LabeledSwitch';
+@import '../components/CheckboxLabeled/CheckboxLabeled';
+@import '../components/RadioButtonLabeled/RadioButtonLabeled';
+@import '../components/SwitchLabeled/SwitchLabeled';
 @import '../components/VerticalListMenu/VerticalListMenu';
 @import '../components/RadioButton/RadioButton';
 @import '../components/RadioGroup/RadioGroup';
@@ -31,7 +31,7 @@
 @import '../components/Switch/Switch';
 @import '../components/Table/Table';
 @import '../components/TextField/TextField';
-@import '../components/ValidatedTextField/ValidatedTextField';
+@import '../components/TextFieldValidated/TextFieldValidated';
 @import '../components/Validation/Validation';
 @import '../components/Tabs/Tabs';
 @import '../components/Overlay/Overlay';
